% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oligo}
\alias{oligo}
\title{Oligodendrocyte array}
\format{
An object of class \code{array} of dimension 300 x 300 x 81 x 4.
}
\usage{
oligo
}
\description{
The oligo dataset in bossrdata consists of a small 4D region-of-interest in cortex from an adult transgenic mouse
expressing EGFP in the cytoplasm of oligodendrocytes (MOBP-EGFP). This same region was imaged at least weekly using two-photon
imaging through a cranial window over the course of cuprizone-treatment and recovery.
}
\keyword{datasets}
